+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <a name="0.13.0">
 
+* can now configure the maximum visible items for 'tags' view via the maxItems option in the field config. The default value is 10.
 * added auto generated id to all fields. Those ids are used in label#for and input#id. You should also reference this id in custom inputs to make the label clickable.
 * added ec-icon component for more flexible icons.